# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Recipe Management API
  description: |
    RESTful API for creating, searching, updating, and deleting recipes.

    **Assumptions**
    - Recipe names are not unique; two recipes may share the same name.
    - Create (`POST`) is not idempotent; each call inserts a new recipe.
    - `includeIngredients` will include all recipe listed in query
    - Ingredient name matching is case-insensitive and exact matching.
    - `instructionSearch` is a case-insensitive substring match (`LIKE %text%`).
    - Update and delete lock the recipe row. Concurrent call will wait up to 3 seconds,
      then receive `409 Conflict`.
    - `X-Trace-Id` is for log tracing only. If not passed, the server generates one
      and returns it on the response.
    - Unit values are not strict, no enums listed. (Good to have)
  version: 0.0.1
  contact:
    name: recipe-service

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Recipes
    description: Recipe CRUD and search

paths:
  /api/v1/recipes:
    get:
      tags: [Recipes]
      operationId: listRecipes
      summary: List recipes
      description: Returns a pagianated recipe list
      parameters:
        - $ref: "#/components/parameters/TraceIdHeader"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - $ref: "#/components/parameters/Sort"
      responses:
        "200":
          description: Page of recipes
          headers:
            X-Trace-Id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipePage"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags: [Recipes]
      operationId: createRecipe
      summary: Create a recipe
      parameters:
        - $ref: "#/components/parameters/TraceIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeRequest"
            example:
              name: Vegetarian Pasta Primavera
              vegetarian: true
              servings: 4
              ingredients:
                - name: pasta
                  quantity: 250
                  unit: grams
                - name: bell peppers
                  quantity: 2
                  unit: pieces
              instructions:
                - stepNumber: 1
                  instructionText: Boil pasta in salted water for 10 minutes.
                - stepNumber: 2
                  instructionText: Sauté chopped bell peppers in olive oil.
      responses:
        "201":
          description: Recipe created
          headers:
            X-Trace-Id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/v1/recipes/search:
    get:
      tags: [Recipes]
      operationId: searchRecipes
      summary: Search recipes
      description: |
        All provided filters are queried with AND, not OR from database
      parameters:
        - $ref: "#/components/parameters/TraceIdHeader"
        - name: vegetarian
          in: query
          required: false
          schema:
            type: boolean
          description: When set, only recipes with this vegetarian flag.
        - name: servings
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
          description: Exact servings count.
        - name: includeIngredients
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
          description: Recipe must include every listed ingredient name (case-insensitive, exact).
        - name: excludeIngredients
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
          description: Recipe must not include any listed ingredient name (case-insensitive, exact).
        - name: instructionSearch
          in: query
          required: false
          schema:
            type: string
          description: Case-insensitive substring match against instruction text.
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - $ref: "#/components/parameters/Sort"
      responses:
        "200":
          description: Page of matching recipes
          headers:
            X-Trace-Id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipePage"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/v1/recipes/{id}:
    parameters:
      - $ref: "#/components/parameters/TraceIdHeader"
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
        description: Recipe id

    get:
      tags: [Recipes]
      operationId: getRecipeById
      summary: Get a recipe by id
      responses:
        "200":
          description: Recipe found
          headers:
            X-Trace-Id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      tags: [Recipes]
      operationId: updateRecipe
      summary: Update a recipe
      description: |
        Update all details of a recipe. Concurrent updat will return `409 Conflict` if the lock cannot be acquired.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeRequest"
      responses:
        "200":
          description: Recipe updated
          headers:
            X-Trace-Id:
              $ref: "#/components/headers/TraceId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecipeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [Recipes]
      operationId: deleteRecipe
      summary: Delete a recipe
      description: |
        Deletes the recipe by id.
        A concurrent update or delete may receive `409 Conflict`.
        A second delete of the same id returns `404`.
      responses:
        "204":
          description: Recipe deleted
          headers:
            X-Trace-Id:
              $ref: "#/components/headers/TraceId"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"

components:
  parameters:
    TraceIdHeader:
      name: X-Trace-Id
      in: header
      required: false
      schema:
        type: string
        maxLength: 128
      description: Optional header, to be included in logs for debugging.
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: One-based page index. `1` is the first page.
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Page size. Values above 100 are capped at 100.
    Sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        default: id,asc
      description: |
        Spring Data sort, e.g. `id,asc` or `name,desc`. Default is `id,asc`.

  headers:
    TraceId:
      description: Optional header, to be included in logs for debugging.
      schema:
        type: string

  responses:
    BadRequest:
      description: Validation failed or the body is malformed.
      headers:
        X-Trace-Id:
          $ref: "#/components/headers/TraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            validation:
              value:
                timestamp: "2026-09-21T23:00:00"
                status: 400
                error: Validation Error
                message: Request validation failed.
                traceId: 11111111-1111-1111-1111-111111111111
                details:
                  name: Recipe name is required
            malformed:
              value:
                timestamp: "2026-09-21T23:00:00"
                status: 400
                error: Bad Request
                message: Request body is missing or malformed.
                traceId: 11111111-1111-1111-1111-111111111111
    NotFound:
      description: Recipe does not exist.
      headers:
        X-Trace-Id:
          $ref: "#/components/headers/TraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            timestamp: "2026-09-21T23:00:00"
            status: 404
            error: Not Found
            message: "Recipe not found with id: 99"
            traceId: 11111111-1111-1111-1111-111111111111
    Conflict:
      description: Concurrent update/delete lock timeout
      headers:
        X-Trace-Id:
          $ref: "#/components/headers/TraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            concurrent:
              value:
                timestamp: "2026-09-21T23:00:00"
                status: 409
                error: Conflict
                message: Recipe was updated concurrently. Retry the request.
                traceId: 11111111-1111-1111-1111-111111111111
            integrity:
              value:
                timestamp: "2026-09-21T23:00:00"
                status: 409
                error: Conflict
                message: Recipe data conflicts with existing records. Retry the request.
                traceId: 11111111-1111-1111-1111-111111111111
    InternalError:
      description: Unexpected server error.
      headers:
        X-Trace-Id:
          $ref: "#/components/headers/TraceId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            timestamp: "2026-09-21T23:00:00"
            status: 500
            error: Internal Server Error
            message: An unexpected error occurred. Retry the request.
            traceId: 11111111-1111-1111-1111-111111111111

  schemas:
    Ingredient:
      type: object
      required: [name, quantity, unit]
      properties:
        name:
          type: string
          minLength: 1
          example: pasta
        quantity:
          type: number
          minimum: 0
          exclusiveMinimum: true
          example: 250
        unit:
          type: string
          minLength: 1
          example: grams

    Instruction:
      type: object
      required: [stepNumber, instructionText]
      properties:
        stepNumber:
          type: integer
          minimum: 1
          example: 1
        instructionText:
          type: string
          minLength: 1
          example: Boil pasta in salted water for 10 minutes.

    RecipeRequest:
      type: object
      required: [name, vegetarian, servings, ingredients, instructions]
      properties:
        name:
          type: string
          minLength: 1
        vegetarian:
          type: boolean
        servings:
          type: integer
          minimum: 1
        ingredients:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/Ingredient"
        instructions:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/Instruction"

    RecipeResponse:
      allOf:
        - type: object
          required: [id]
          properties:
            id:
              type: integer
              format: int64
              example: 1
        - $ref: "#/components/schemas/RecipeRequest"

    RecipePage:
      type: object
      properties:
        recipes:
          type: array
          items:
            $ref: "#/components/schemas/RecipeResponse"
        totalCount:
          type: integer
          format: int64
          description: Total number of recipes matching the request.
        page:
          type: integer
          description: Current page index, 1 is 1st page

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        traceId:
          type: string
          nullable: true
        details:
          type: object
          additionalProperties:
            type: string
          description: Present on validation errors; field name to message.